tests/assert-clean.sh: Test for dirty and untracked files independently

+++ tests/assert-clean.sh
@@ -6,11 +6,6 @@ if [ ! -e .git ]; then
 	exit
 fi
 
-if [ -n "$(git status -s)" ]; then
-	echo "Repository is already dirty, skipping \"$0\"" >&2
-	exit
-fi
-
 
 . tools/lib.sh
 
@@ -21,21 +16,32 @@ cp -pR ./ "$R"
 cd "$R"
 
 
+DIRTY=false
+UNTRACKED=false
+
+if [ -n "$(git status -s)" ]; then
+	DIRTY=true
+fi
+
+if [ -n "$(git clean -nffdx)" ]; then
+	UNTRACKED=true
+fi
+
 {
 	make -s clean
 
 	printf '%s: "clean" target deletes all derived assets...' \
 		"$(yellow "$0")"
 
-	if [ -n "$(git status -s)" ]; then
-		printf ' ERR.\n'
+	if [ "$DIRTY" = false ] && [ -n "$(git status -s)" ]; then
+		printf ' %s.\n' "$(red 'ERR')"
 		echo 'Repository left dirty:'
 		git status
 		exit 1
 	fi
 
-	if [ -n "$(git clean -nffdx)" ]; then
-		printf ' ERR.\n'
+	if [ "$UNTRACKED" = false ] && [ -n "$(git clean -nffdx)" ]; then
+		printf ' %s.\n' "$(red 'ERR')"
 		echo 'Untracked files left:'
 		git clean -ffdx --dry-run
 		exit 1
